Fixes Bug where opening Specify when the Wizard is still open on the last dialog. Also, fixes the reoding of deleting the agent_disciplines table
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@8923 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/dbsupport/SpecifySchemaUpdateService.java
@@ -1957,7 +1957,10 @@ public class SpecifySchemaUpdateService extends SchemaUpdateService
                     } catch (Exception ex) {}
                 }
                 
-                BasicSQLUtils.update(conn, "DROP TABLE agent_discipline");
+                if (BasicSQLUtils.doesTableExist(conn, "agent_discipline"))
+                {
+                    BasicSQLUtils.update(conn, "DROP TABLE agent_discipline");
+                }
                 
                 // Add New Fields to Address
                 tblName = setTableTitleForFrame(Agent.getClassTableId());
